+++ src/sbin/isakmpd/connection.c
@@ -1,3 +1,4 @@
+/* $MirOS: src/sbin/isakmpd/connection.c,v 1.3 2005/04/26 15:42:37 tg Exp $ */
 /* $OpenBSD: connection.c,v 1.32 2005/05/04 10:05:01 hshoexer Exp $	 */
 /* $EOM: connection.c,v 1.28 2000/11/23 12:21:18 niklas Exp $	 */
 
@@ -408,9 +409,9 @@ connection_report(void)
 	for (conn = TAILQ_FIRST(&connections); conn;
 	    conn = TAILQ_NEXT(conn, link))
 		LOG_DBG((LOG_REPORT, 0,
-		    "connection_report: connection %s next check %ld seconds",
+		    "connection_report: connection %s next check %lld seconds",
 		    (conn->name ? conn->name : "<unnamed>"),
-		    conn->ev->expiration.tv_sec - now.tv_sec));
+		    (int64_t)(conn->ev->expiration.tv_sec - now.tv_sec)));
 	for (pconn = TAILQ_FIRST(&connections_passive); pconn;
 	    pconn = TAILQ_NEXT(pconn, link))
 		LOG_DBG((LOG_REPORT, 0,
