+++ src/sys/dev/pcmcia/if_sm_pcmcia.c
@@ -1,3 +1,4 @@
+/**	$MirOS: src/sys/dev/pcmcia/if_sm_pcmcia.c,v 1.2 2005/03/06 21:27:54 tg Exp $ */
 /*	$OpenBSD: if_sm_pcmcia.c,v 1.23 2005/06/08 17:03:01 henning Exp $	*/
 /*	$NetBSD: if_sm_pcmcia.c,v 1.11 1998/08/15 20:47:32 thorpej Exp $  */
 
@@ -184,7 +185,7 @@ sm_pcmcia_attach(parent, self, aux)
 		return;
 	}
 
-	printf(" port 0x%lx/%d", psc->sc_pcioh.addr, psc->sc_pcioh.size);
+	printf(" port 0x%lx/%d", psc->sc_pcioh.addr, (int)(psc->sc_pcioh.size));
 
 	/*
 	 * First try to get the Ethernet address from FUNCE/LANNID tuple.
